refactor: Reorganize Torii and error memory into runtime and exec
