+++ internal/agents/exec/toolCall.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 	"sync"
 
-	allowTool "github.com/pardnchiu/agenvoy/internal/agents/exec/allow/tool"
+	"github.com/pardnchiu/agenvoy/internal/agents/exec/allow/tool"
+	"github.com/pardnchiu/agenvoy/internal/agents/exec/memory"
 	agentTypes "github.com/pardnchiu/agenvoy/internal/agents/types"
-	"github.com/pardnchiu/agenvoy/internal/filesystem/errorMemory"
-	"github.com/pardnchiu/agenvoy/internal/filesystem/errorMemory/toolError"
+	"github.com/pardnchiu/agenvoy/internal/filesystem/toolError"
 	"github.com/pardnchiu/agenvoy/internal/runtime"
 	"github.com/pardnchiu/agenvoy/internal/tools"
 	toolRegister "github.com/pardnchiu/agenvoy/internal/tools/register"
@@ -217,7 +217,7 @@ func toolCall(ctx context.Context, exec *toolTypes.Executor, choice agentTypes.O
 			if toolFailCount[s.hash] >= MaxRetry {
 				result = fmt.Sprintf("[ABORT] tool=%s 連續 %d 次失敗: %s\n請改用其他工具或顯著調整參數，不要使用相同工具 %s。", s.name, toolFailCount[s.hash], s.execErr, s.name)
 			} else {
-				if hint := errorMemory.Search(ctx, s.name, s.execErr, 3); hint != "" {
+				if hint := memory.Search(ctx, s.name, s.execErr, 3); hint != "" {
 					result = fmt.Sprintf("[RETRY_REQUIRED] tool=%s failed: %s\nrelated_errors: %s\nFix the arguments and call %s again immediately. Do NOT output this message as your response.", s.name, s.execErr, hint, s.name)
 				} else {
 					result = fmt.Sprintf("[RETRY_REQUIRED] tool=%s failed: %s\nFix the arguments and call %s again immediately. Do NOT output this message as your response.", s.name, s.execErr, s.name)
@@ -225,7 +225,7 @@ func toolCall(ctx context.Context, exec *toolTypes.Executor, choice agentTypes.O
 				delete(alreadyCall, s.hash)
 			}
 		} else if result == "" || result == "no data" {
-			if hint := errorMemory.Search(ctx, s.name, "no data", 3); hint != "" {
+			if hint := memory.Search(ctx, s.name, "no data", 3); hint != "" {
 				result = hint
 			} else {
 				result = "no data"
