+++ internal/filesystem/filesystem.go
@@ -130,12 +130,3 @@ func PagePath(sessionID string) string {
 	return filepath.Join(SessionsDir, sessionID, "page")
 }
 
-const AllowListRelPath = ".agenvoy/allow_list"
-
-func AllowListPath(workDir string) string {
-	return filepath.Join(workDir, AllowListRelPath)
-}
-
-func AllowListDir(workDir string) string {
-	return filepath.Join(workDir, filepath.Dir(AllowListRelPath))
-}
