+++ internal/filesystem/toolError/toolError.go
